[+] Hot edit bug fixed. [+] Introduced more bugs to fix later 🤣.
