+++ lib/widgets/target_node.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+import '../models/tree_node.dart';
 import '../providers/inspector_provider.dart';
 import '../models/node.dart';
 import '../models/registerer.dart';
